pyproject.toml
[project]
name = "api-gateway-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "fastapi[standard]>=0.121.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.4",
]
